@@ -26,7 +26,7 @@ const CharInfo = (props) => {
         if (!charId) {
             return; 
         } 
-           clearError()
+           
             getCharacter(charId)
                 .then(onCharLoaded)       
         }
