@@ -1,4 +1,5 @@
 import {Component} from 'react';
+import ErrorMessage from "../errorMessage/ErrorMessage";
 
 class ErrorBoundary extends Component {
     state = {
@@ -13,14 +14,25 @@ class ErrorBoundary extends Component {
         })
     }
     
+    // render() {
+    // if (this.state.error) {
+    //     return <h2>Something went wrong</h2>
+    // } 
+    //  return (this.props.children);
+    // }
+
+    
     render() {
-    if (this.state.error) {
-        return <h2>Something went wrong</h2>
-    } 
+        if (this.state.error) {
+            return <ErrorMessage/>
+        }
 
-
-     return (this.props.children);
+        return this.props.children;
     }
+
+
+
+
 
 }
 
